chore: sisyfos: move intenal types away from the types-lib
