+++ src/devices/sisyfosAPI.ts
@@ -1,20 +1,11 @@
 import * as osc from 'osc'
-import {
-	SisyfosState,
-	SisyfosCommand,
-	Commands,
-	ValueCommand,
-	StringCommand,
-	SisyfosAPIState,
-	SisyfosChannel
-} from '../types/src/sisyfos'
 import { EventEmitter } from 'events'
 
 /** How often to check connection status */
 const CONNECTIVITY_INTERVAL = 3000 // ms
 const CONNECTIVITY_TIMEOUT = 1000 // ms
 
-export class SisyfosInterface extends EventEmitter {
+export class SisyfosApi extends EventEmitter {
 	host: string
 	port: number
 
@@ -75,31 +66,31 @@ export class SisyfosInterface extends EventEmitter {
 	}
 
 	send (command: SisyfosCommand) {
-		if (command.type === Commands.TAKE) {
+		if (command.type === SisyfosCommandType.TAKE) {
 			this._oscClient.send({ address: '/take', args: [] })
-		} else if (command.type === Commands.CLEAR_PST_ROW) {
+		} else if (command.type === SisyfosCommandType.CLEAR_PST_ROW) {
 			this._oscClient.send({ address: '/clearpst', args: [] })
-		} else if (command.type === Commands.LABEL) {
+		} else if (command.type === SisyfosCommandType.LABEL) {
 			this._oscClient.send({ address: `/ch/${(command as StringCommand).channel + 1}/label`, args: [{
 				type: 's',
 				value: (command as StringCommand).value
 			}] })
-		} else if (command.type === Commands.TOGGLE_PGM) {
+		} else if (command.type === SisyfosCommandType.TOGGLE_PGM) {
 			this._oscClient.send({ address: `/ch/${(command as ValueCommand).channel + 1}/pgm`, args: [{
 				type: 'i',
 				value: (command as ValueCommand).value
 			}] })
-		} else if (command.type === Commands.TOGGLE_PST) {
+		} else if (command.type === SisyfosCommandType.TOGGLE_PST) {
 			this._oscClient.send({ address: `/ch/${(command as ValueCommand).channel + 1}/pst`, args: [{
 				type: 'i',
 				value: (command as ValueCommand).value
 			}] })
-		} else if (command.type === Commands.SET_FADER) {
+		} else if (command.type === SisyfosCommandType.SET_FADER) {
 			this._oscClient.send({ address: `/ch/${(command as ValueCommand).channel + 1}/faderlevel`, args: [{
 				type: 'f',
 				value: (command as ValueCommand).value
 			}] })
-		} else if (command.type === Commands.VISIBLE) {
+		} else if (command.type === SisyfosCommandType.VISIBLE) {
 			this._oscClient.send({ address: `/ch/${(command as ValueCommand).channel + 1}/visible`, args: [{
 				type: 'i',
 				value: (command as ValueCommand).value
@@ -240,5 +231,70 @@ export class SisyfosInterface extends EventEmitter {
 		})
 
 		return deviceState
+	}
+}
+
+export enum SisyfosCommandType {
+	TOGGLE_PGM = 'togglePgm',
+	TOGGLE_PST = 'togglePst',
+	SET_FADER = 'setFader',
+	CLEAR_PST_ROW = 'clearPstRow',
+	LABEL = 'label',
+	TAKE = 'take',
+	VISIBLE = 'visible',
+	RESYNC = 'resync'
+}
+
+export interface BaseCommand {
+	type: SisyfosCommandType
+}
+
+export interface ChannelCommand {
+	type: SisyfosCommandType.SET_FADER | SisyfosCommandType.TOGGLE_PGM | SisyfosCommandType.TOGGLE_PST | SisyfosCommandType.LABEL | SisyfosCommandType.VISIBLE
+	channel: number
+	value: boolean | number | string
+}
+
+export interface BoolCommand extends ChannelCommand {
+	type: SisyfosCommandType.VISIBLE
+	value: boolean
+}
+export interface ValueCommand extends ChannelCommand {
+	type: SisyfosCommandType.TOGGLE_PGM | SisyfosCommandType.TOGGLE_PST | SisyfosCommandType.SET_FADER
+	value: number
+}
+
+export interface StringCommand extends ChannelCommand {
+	type: SisyfosCommandType.LABEL
+	value: string
+}
+
+export interface ResyncCommand extends BaseCommand {
+	type: SisyfosCommandType.RESYNC
+}
+
+export type SisyfosCommand = BaseCommand | ValueCommand | BoolCommand | StringCommand | ResyncCommand
+
+export interface SisyfosChannel extends SisyfosAPIChannel {
+	tlObjIds: string[]
+}
+export interface SisyfosState {
+	channels: { [index: string]: SisyfosChannel }
+	resync: boolean
+}
+
+// ------------------------------------------------------
+// Interfaces for the data that comes over OSC:
+export interface SisyfosAPIChannel {
+	faderLevel: number
+	pgmOn: number
+	pstOn: number
+	label: string
+	visible: boolean
+}
+
+export interface SisyfosAPIState {
+	channels: {
+		[index: string]: SisyfosAPIChannel
 	}
 }
